Progress on making DbTestCase obsolete.
