+++ module/VuFind/tests/integration-tests/src/VuFindTest/Auth/ShibbolethTest.php
@@ -42,8 +42,9 @@ use VuFind\Auth\Shibboleth\SingleIdPConfigurationLoader;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org Main Page
  */
-class ShibbolethTest extends \VuFindTest\Unit\DbTestCase
+class ShibbolethTest extends \VuFindTest\Unit\TestCase
 {
+    use \VuFindTest\Unit\LiveDatabaseTrait;
     use \VuFindTest\Unit\UserCreationTrait;
 
     protected $user1 = [
@@ -122,8 +123,7 @@ class ShibbolethTest extends \VuFindTest\Unit\DbTestCase
         }
         $obj = new Shibboleth($this->createMock(\Laminas\Session\ManagerInterface::class), $loader,
             $this->createMock(\Laminas\Http\PhpEnvironment\Request::class));
-        $initializer = new \VuFind\ServiceManager\ServiceInitializer();
-        $initializer($this->getServiceManager(), $obj);
+        $obj->setDbTableManager($this->getLiveTableManager());
         $obj->setConfig($config);
         return $obj;
     }
